Implemented toaster functionality.
The toaster can be activated when the two toasts are in it and the cable is plugged in.

Fixed problem with the coffee and the milk: if you put twice the milk in the cup, then it's filled with only milk, not coffee with milk.

Tested and succeeded app in Android.